apps/backend/ml/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
lightfm
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
